build: Keep comments in builds
Keeping comments has the benefit of keeping the JSDoc, which some our of
code has, and is lost for any consumer of the packages, because they are
stripped during build.

An alternative to this is to build the JS code in a first step, with no
comments, and then generate TS declarations in a second step while
keeping them. I do not think it's worth it, though, this library doesn't
really have in-code comments.